Upgrade to SFML3 (#41)
* Updated for SFML 3

* Removed hard-coded paths from CMakeLists.txt.

* Update README.md

* Combined redundant VertexArrays into single VertexArray.

* Updated CMakeLists.txt to a more modern approach.

* Upgraded demo.cpp for SFML3

* Upped uint16s to uint32s to prevent heap corruption

* Upped more uint16s to uint32s to prevent heap corruption

* Fixed cmakelists not including SFML::System for demo

* Fixed LightingArea's VertexArrays being initialized with size of 3 instead of 6 (one triangle vs. two)

+++ src/Color.cpp
@@ -2,15 +2,15 @@
 
 namespace sfu{
     sf::Color darken(const sf::Color& c, float r){
-        return sf::Color(c.r * (1.f-r), c.g * (1.f-r), c.b * (1.f-r), c.a);
+        return sf::Color(uint8_t(c.r * (1.f-r)), uint8_t(c.g * (1.f-r)), uint8_t(c.b * (1.f-r)), c.a);
     }
     
     sf::Color lighten(const sf::Color& c, float r){
-        return sf::Color(c.r * (1.f+r), c.g * (1.f+r), c.b * (1.f+r), c.a);
+        return sf::Color(uint8_t(c.r * (1.f+r)), uint8_t(c.g * (1.f+r)), uint8_t(c.b * (1.f+r)), c.a);
     }
     
     sf::Color interpolate(const sf::Color& c1, const sf::Color& c2, float r){
-        return sf::Color(c1.r + (c2.r - c1.r) * r, c1.g + (c2.g - c1.g) * r, c1.b + (c2.b - c1.b) * r, c1.a + (c2.a - c1.a) * r);
+        return sf::Color(uint8_t(c1.r + (c2.r - c1.r) * r), uint8_t(c1.g + (c2.g - c1.g) * r), uint8_t(c1.b + (c2.b - c1.b) * r), uint8_t(c1.a + (c2.a - c1.a) * r));
     }
     
     sf::Color complementary(const sf::Color& c){
